fixed angle check in squares.py (thanks to Abid Rahman)

+++ samples/python2/squares.py
@@ -3,7 +3,7 @@ import cv2
 
 
 def angle_cos(p0, p1, p2):
-    d1, d2 = p0-p1, p2-p1
+    d1, d2 = (p0-p1).astype('float'), (p2-p1).astype('float')
     return abs( np.dot(d1, d2) / np.sqrt( np.dot(d1, d1)*np.dot(d2, d2) ) )
 
 def find_squares(img):
